+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Version 4.0
 - ``setuptools_scm`` is included by default in ``setup.cfg`` and ``setup.py``
 - API changed to use ``pyscaffold.operations`` instead of integer flags, "issue" #271
 - Allow ``string.Template`` and ``callable`` as file contents in project structure, "proposal" #295
+- Extract file system functions from ``utils.py`` into ``file_system.py``
 
 
 Current versions